+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Upgraded scanoss.js SDK version to v0.27.0
 - Improved API endpoint deletion to check both URL and API KEY for unique identification
+- Increased default request chunk limits
 
 ## [1.25.0] - 2025-11-05
 ### Added